refactor: API 호출 관련 로직을 domain 폴더 하위로 이동
